Move cached_cumsum to dask/array/slicing.py
It can't go in dask/utils.py because the top level is not supposed to
depend on numpy.
