+++ dask/array/tests/test_slicing.py
@@ -10,7 +10,7 @@ import dask
 import dask.array as da
 from dask.array.slicing import (_sanitize_index_element, _slice_1d,
                                 new_blockdim, sanitize_index, slice_array,
-                                take, normalize_index, slicing_plan)
+                                take, normalize_index, slicing_plan, cached_cumsum)
 from dask.array.utils import assert_eq, same_keys
 
 
@@ -797,6 +797,33 @@ def test_pathological_unsorted_slicing():
 
     assert '10' in str(info.list[0])
     assert 'out-of-order' in str(info.list[0])
+
+
+def test_cached_cumsum():
+    a = (1, 2, 3, 4)
+    x = cached_cumsum(a)
+    y = cached_cumsum(a, initial_zero=True)
+    np.testing.assert_array_equal(x, [1, 3, 6, 10])
+    assert x.dtype == np.int64
+    np.testing.assert_array_equal(y, [0, 1, 3, 6, 10])
+    assert y.dtype == np.int64
+
+
+def test_cached_cumsum_nan():
+    a = (1, np.nan, 3)
+    x = cached_cumsum(a)
+    y = cached_cumsum(a, initial_zero=True)
+    np.testing.assert_array_equal(x, [1, np.nan, np.nan])
+    assert x.dtype == np.float64
+    np.testing.assert_array_equal(y, [0, 1, np.nan, np.nan])
+    assert y.dtype == np.float64
+
+
+def test_cached_cumsum_non_tuple():
+    a = [1, 2, 3]
+    np.testing.assert_array_equal(cached_cumsum(a), [1, 3, 6])
+    a[1] = 4
+    np.testing.assert_array_equal(cached_cumsum(a), [1, 5, 8])
 
 
 @pytest.mark.parametrize('params', [(2, 2, 1), (5, 3, 2)])
